Extract ContextData to separate file
